feat: better organize integration tests
